MaterialModelsTesting

Documentation for MaterialModelsTesting.

MaterialModelsTesting.are_derivatives_equal

MaterialModelsTesting.compare_derivatives

MaterialModelsTesting.obtain_numerical_material_derivative!

MaterialModelsTesting.obtain_numerical_material_derivative!

MaterialModelsTesting.runstrain

MaterialModelsTesting.runstrain_diff

MaterialModelsTesting.runstresstate

MaterialModelsTesting.runstresstate_diff

MaterialModelsTesting.test_derivative

MaterialModelsTesting.test_material

MaterialModelsTesting.test_vectorconversion

MaterialModelsTesting.are_derivatives_equal — Method

are_derivatives_equal(args...; maxtol = Inf, kwargs...)

Calls compare_derivatives (forwarding args and kwargs) and checks if the derivatives are equal. The maximum allowed absolute tolerance can also be checked by setting maxtol.

MaterialModelsTesting.compare_derivatives — Method

compare_derivatives(dydx::Matrix, dydx_num::Matrix, y::Vector, Δx::Vector; tolscale = 1)

Comparing derivatives with numerical values can be tricky, due to catastrophic cancellation easily making the numerical derivatives inaccurate. This function uses the following formula to determine the appropriate tolerance for comparing the derivatives, which is based on the values of the differentiated function, 'y', the pertubation for each entry, Δx. 'atol = 2 * eps(y) / Δx`.

It returns the error scaled by the tolerance for all entries, e, such that the derivatives can be considered equal if all(x -> x ≤ 1, e)

MaterialModelsTesting.obtain_numerical_material_derivative! — Method

obtain_numerical_material_derivative!(deriv, m, ϵ, old, Δt; fdtype = Val{:forward}, kwargs...)

Obtain the numerical derivative of the material m at the strain ϵ and old state variables, old, for a time step Δt. fdtype and kwargs... are passed to FiniteDiff.finite_difference_jacobian.

MaterialModelsTesting.obtain_numerical_material_derivative! — Method

obtain_numerical_material_derivative!(ssd, stress_state, m, ϵ, old, Δt; fdtype = Val{:forward}, kwargs...)

Obtain the numerical derivative of the material m considering the stress_state iterations at the strain ϵ and old state variables, old, for a time step Δt. fdtype and kwargs... are passed to FiniteDiff.finite_difference_jacobian.

MaterialModelsTesting.runstrain — Method

runstrain(m, TensorBase, ϵ_end::Union{Number, AbstractTensor}, ij::NTuple{2, Int}, t_end, num_steps)

Simulate fully strain-controlled loading from zero to ϵ_end strain, and return σ[i,j]. If ϵ_end::Number is passed, loading is applied to the strain ϵ_end eᵢ⊗eⱼ. TensorBase should be SymmetricTensor{2,3} for small strain and Tensor{2,3} for finite strain. For finite strain, the deformation gradient I2 + ϵ_end eᵢ⊗eⱼ is applied.

Returns the vector of stresses as well as the final state.

MaterialModelsTesting.runstrain_diff — Method

runstrain_diff(m, TensorBase, ϵ_end::Union{Number, AbstractTensor}, ij::NTuple{2, Int}, t_end, num_steps)

Simulate fully strain-controlled loading from zero to ϵ_end strain, and return σ[i,j] and its derivatives wrt. to the material parameters. If ϵ_end::Number is passed, loading is applied to the strain ϵ_end eᵢ⊗eⱼ. TensorBase should be SymmetricTensor{2,3} for small strain and Tensor{2,3} for finite strain. For finite strain, the deformation gradient I2 + ϵ_end eᵢ⊗eⱼ is applied.

Returns the vector of stresses, the final state, the derivatives dσᵢⱼ/dp for all time steps, and the final derivatives diff::MaterialDerivatives.

MaterialModelsTesting.runstresstate — Method

runstresstate(stress_state, m, TensorBase, ϵ_end::Union{Number, AbstractTensor}, ij::NTuple{2, Int}, t_end, num_steps)

Simulate the stress_state from zero to ϵ_end strain, and return σ[i,j]. If ϵ_end::Number is passed, loading is applied to the strain ϵ_end eᵢ⊗eⱼ. TensorBase should be SymmetricTensor{2,3} for small strain and Tensor{2,3} for finite strain. For finite strain, the deformation gradient I2 + ϵ_end eᵢ⊗eⱼ is applied.

Returns the vector of stresses and the final state variables.

MaterialModelsTesting.runstresstate_diff — Method

runstresstate_diff(stress_state, m, TensorBase, ϵ_end::Union{Number, AbstractTensor}, ij::NTuple{2, Int}, t_end, num_steps)

Simulate the stress_state from zero to ϵ_end strain, and return σ[i,j] and its derivatives wrt. to the material parameters. If ϵ_end::Number is passed, loading is applied to the strain ϵ_end eᵢ⊗eⱼ.

Returns the vector of stresses, the final state, the derivatives dσᵢⱼ/dp for all time steps, and the final derivatives diff::StressStateDerivatives.

MaterialModelsTesting.test_derivative — Method

test_derivative(m, ϵ, state, Δt; comparesettings = (), numdiffsettings = (), diff = MaterialDerivatives(m))

This function is used to compare the analytically implemented derivative (calling differentiate_material!(diff, m, ϵ, args...)) with the result from obtain_numerical_material_derivative!(diff, m, ϵ, args...; numdiffsettings...).

diff::MaterialDerivatives can be passed to check that the old value of diff.dsdp is correctly accounted for.

comparesettings::NamedTuple are passed as kwargs to compare_derivatives, which compares the two matrices. Please see its docstring for further details.

numdiffsettings::NamedTuple are passed as kwargs to obtain_numerical_material_derivative!, please see its docstring for further details.

MaterialModelsTesting.test_material — Function

test_material(m::AbstractMaterial, [loadinfo::LoadInfo];
    basics = true, conversions = false, differentiation = false
    )

Function to quickly test that the material m follows the MaterialModelsBase interface, and that what generally should hold true, does. All material models should pass the basics tests, which are on by default, whereas the conversions and differentiation tests are optional.

loadinfo gives information about how much the material is loaded, and different values can be supplied if required for the specific material model/parameters.

basics (default: true)

Tests that required functions - initial_material_state and allocate_material_cache returns the correct types, and runs the material_response to check that the outputs are of correct type. Furthermore, it checks that the tangent stiffness matches with the numerical derivative, although this test is only approximate.

conversions (default: false)

Tests the conversions tovector, tovector!, fromvector for both the material parameters and the state variables, along with the associated methods get_vector_length, get_num_statevars, and get_vector_eltype. Furthermore, it is tested that initial_material_state returns a state using the element type from the material, and that the conversions respect the different element types of the given vector. Since also Float32 is used in this testing, it is important not to default to using Float64 when creating e.g. state variables, i.e. use zero(Tensor{2,3,T}) instead of zeros(Tensor{2,3}).

differentiation (default: false)

Tests that the differentiate_material! function works correctly. Uses BigFloat to calculate the numerical derivatives to avoid floating point errors.

MaterialModelsTesting.test_vectorconversion — Method

test_vectorconversion(T::Type, obj)

Test conversions using tovector!, tovector, and fromvector of the object obj, for the number type T in the vector (which may be different from that stored in obj).